server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "bazi-analysis FastAPI backend (foundation layer)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29",
    "alembic>=1.13",
    "cryptography>=42.0",
    "structlog>=24.1",
    "httpx>=0.27",
    "openai>=1.40",
    "paipan",
    "pillow>=10",
    "zhconv>=1.4",
    "colorthief>=0.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-xdist>=3.5",
    "testcontainers[postgres]>=4.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.uv.sources]
paipan = { workspace = true }

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
